pvnet/requirements.txt
cython==0.29.24
yacs==0.1.8
opencv-python==4.5.4.58
tqdm==4.62.3
pycocotools==2.0.2
matplotlib==3.4.3
plyfile==0.7.4
scikit-image==0.18.3
scikit-learn
PyOpenGL==3.1.5
ipdb==0.13.9
cyglfw3==3.1.0.2
pyassimp==4.1.4
progressbar==2.5
open3d-python==0.3.0.0
tensorflow==2.8.0
tensorboardX==2.4
cffi==1.15.0

